backend/requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
beautifulsoup4==4.14.2
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
Django==5.2.8
django-cors-headers==4.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.1
faiss-cpu==1.13.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
groq==0.36.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
murmurhash==1.0.15
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
preshed==3.0.12
prompt_toolkit==3.0.52
psutil==7.1.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0+cpu
torchaudio==2.9.0+cpu
torchvision==0.24.0+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
wheel==0.45.1
wrapt==2.0.1
